feat: creates StorageService for s3 paths
